+++ assets/components/src/action-card/index.js
@@ -7,7 +7,7 @@
  */
 import { Component } from '@wordpress/element';
 import { ExternalLink, ToggleControl } from '@wordpress/components';
-import { Icon, check } from '@wordpress/icons';
+import { Icon, check, chevronDown, chevronUp } from '@wordpress/icons';
 
 /**
  * Internal dependencies
@@ -21,12 +21,21 @@ import './style.scss';
 import classnames from 'classnames';
 
 class ActionCard extends Component {
-	backgroundImageStyles = url => {
-		return url ? { backgroundImage: `url(${ url })` } : {};
+	state = {
+		expanded: false,
 	};
 
 	/**
-	 * Render
+	 * When the collapse prop is updated to true, collapse the card if already expanded.
+	 */
+	componentDidUpdate( prevProps ) {
+		if ( ! prevProps.collapse && this.props.collapse && this.state.expanded ) {
+			this.setState( { expanded: false } );
+		}
+	}
+
+	/**
+	 * Render.
 	 */
 	render() {
 		const {
@@ -60,19 +69,31 @@ class ActionCard extends Component {
 			toggleChecked,
 			toggleOnChange,
 			hasGreyHeader,
+			hasWhiteHeader,
 			isPending,
+			expandable = false,
 		} = this.props;
+
+		const { expanded } = this.state;
+
 		const hasChildren = notification || children;
 		const classes = classnames(
 			'newspack-action-card',
 			simple && 'newspack-card--is-clickable',
 			hasGreyHeader && 'newspack-card--has-grey-header',
+			hasWhiteHeader && 'newspack-card--has-white-header',
 			hasChildren && 'newspack-card--has-children',
 			indent && 'newspack-card--indent',
 			isSmall && 'is-small',
 			isMedium && 'is-medium',
+			checkbox && 'has-checkbox',
+			expandable && 'is-expandable',
+			actionContent && 'has-action-content',
 			className
 		);
+		const backgroundImageStyles = url => {
+			return url ? { backgroundImage: `url(${ url })` } : {};
+		};
 		const titleProps =
 			toggleOnChange && ! titleLink && ! disabled
 				? { onClick: () => toggleOnChange( ! toggleChecked ), tabIndex: '0' }
@@ -95,7 +116,7 @@ class ActionCard extends Component {
 							<a href={ imageLink }>
 								<div
 									className="newspack-action-card__image"
-									style={ this.backgroundImageStyles( image ) }
+									style={ backgroundImageStyles( image ) }
 								/>
 							</a>
 						</div>
@@ -118,7 +139,13 @@ class ActionCard extends Component {
 						<Grid columns={ 1 } gutter={ 8 } noMargin>
 							<h2>
 								<span className="newspack-action-card__title" { ...titleProps }>
-									{ titleLink ? <a href={ titleLink }>{ title }</a> : title }
+									{ titleLink && <a href={ titleLink }>{ title }</a> }
+									{ ! titleLink && expandable && (
+										<Button isLink onClick={ () => this.setState( { expanded: ! expanded } ) }>
+											{ title }
+										</Button>
+									) }
+									{ ! titleLink && ! expandable && title }
 								</span>
 								{ badges?.length &&
 									badges.map( ( badgeText, i ) => (
@@ -135,7 +162,7 @@ class ActionCard extends Component {
 							) }
 						</Grid>
 					</div>
-					{ ( actionText || isDisplayingSecondaryAction || actionContent ) && (
+					{ ! expandable && ( actionText || isDisplayingSecondaryAction || actionContent ) && (
 						<div className="newspack-action-card__region newspack-action-card__region-right">
 							{ /* eslint-disable no-nested-ternary */ }
 							{ actionContent && actionContent }
@@ -176,6 +203,11 @@ class ActionCard extends Component {
 							) }
 						</div>
 					) }
+					{ expandable && (
+						<Button onClick={ () => this.setState( { expanded: ! expanded } ) }>
+							<Icon icon={ expanded ? chevronUp : chevronDown } height={ 24 } width={ 24 } />
+						</Button>
+					) }
 				</div>
 				{ notification && (
 					<div className="newspack-action-card__notification newspack-action-card__region-children">
@@ -193,7 +225,9 @@ class ActionCard extends Component {
 						) }
 					</div>
 				) }
-				{ children && <div className="newspack-action-card__region-children">{ children }</div> }
+				{ children && ( ( expandable && expanded ) || ! expandable ) && (
+					<div className="newspack-action-card__region-children">{ children }</div>
+				) }
 			</Card>
 		);
 	}