+++ includes/plugins/wc-memberships/class-wc-memberships.php
@@ -31,11 +31,16 @@ class WC_Memberships {
 		add_action( 'init', [ __CLASS__, 'register_meta' ] );
 		add_action( 'admin_init', [ __CLASS__, 'redirect_cpt' ] );
 		add_action( 'admin_init', [ __CLASS__, 'handle_edit_gate' ] );
+		add_action( 'wp_enqueue_scripts', [ __CLASS__, 'enqueue_scripts' ] );
 		add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'enqueue_block_editor_assets' ] );
-		add_filter( 'wc_memberships_notice_html', [ __CLASS__, 'notice_html' ], 100 );
+		add_filter( 'wc_memberships_notice_html', [ __CLASS__, 'notice_html' ], 100, 4 );
 		add_filter( 'wc_memberships_restricted_content_excerpt', [ __CLASS__, 'excerpt' ], 100, 3 );
+		add_filter( 'wc_memberships_message_excerpt_apply_the_content_filter', '__return_false' );
+		add_action( 'wp_footer', [ __CLASS__, 'render_overlay_gate' ], 1 );
 		add_action( 'wp_footer', [ __CLASS__, 'render_js' ] );
 		add_filter( 'newspack_popups_assess_has_disabled_popups', [ __CLASS__, 'disable_popups' ] );
+
+		include __DIR__ . '/class-block-patterns.php';
 	}
 
 	/**
@@ -119,6 +124,28 @@ class WC_Memberships {
 				'single'         => true,
 			]
 		);
+		\register_meta(
+			'post',
+			'overlay_position',
+			[
+				'object_subtype' => self::GATE_CPT,
+				'show_in_rest'   => true,
+				'type'           => 'string',
+				'default'        => 'center',
+				'single'         => true,
+			]
+		);
+		\register_meta(
+			'post',
+			'overlay_size',
+			[
+				'object_subtype' => self::GATE_CPT,
+				'show_in_rest'   => true,
+				'type'           => 'string',
+				'default'        => 'medium',
+				'single'         => true,
+			]
+		);
 	}
 
 	/**
@@ -132,6 +159,37 @@ class WC_Memberships {
 		}
 	}
 
+	/**
+	 * Enqueue frontend scripts and styles for gated content.
+	 */
+	public static function enqueue_scripts() {
+		if ( ! self::has_gate() ) {
+			return;
+		}
+		if ( ! is_singular() || ! self::is_post_restricted() ) {
+			return;
+		}
+		$gate_post_id = self::get_gate_post_id();
+		$style        = \get_post_meta( $gate_post_id, 'style', true );
+		if ( 'overlay' !== $style ) {
+			return;
+		}
+		$handle = 'newspack-memberships-gate-overlay';
+		\wp_enqueue_script(
+			$handle,
+			Newspack::plugin_url() . '/dist/memberships-gate-overlay.js',
+			[],
+			filemtime( dirname( NEWSPACK_PLUGIN_FILE ) . '/dist/memberships-gate-overlay.js' ),
+			true
+		);
+		\wp_script_add_data( $handle, 'async', true );
+		\wp_enqueue_style(
+			$handle,
+			Newspack::plugin_url() . '/dist/memberships-gate-overlay.css',
+			[],
+			filemtime( dirname( NEWSPACK_PLUGIN_FILE ) . '/dist/memberships-gate-overlay.css' )
+		);
+	}
 
 	/**
 	 * Enqueue block editor assets.
@@ -209,7 +267,7 @@ class WC_Memberships {
 		if ( ! $post_id ) {
 			$post_id = get_the_ID();
 		}
-		if ( ! function_exists( 'wc_memberships_is_post_content_restricted' ) || ! wc_memberships_is_post_content_restricted( $post_id ) ) {
+		if ( ! function_exists( 'wc_memberships_is_post_content_restricted' ) || ! \wc_memberships_is_post_content_restricted( $post_id ) ) {
 			return false;
 		}
 		return ! is_user_logged_in() || ! current_user_can( 'wc_memberships_view_restricted_post_content', $post_id );
@@ -267,11 +325,18 @@ class WC_Memberships {
 	 * Filter the notice HTML.
 	 *
 	 * @param string $notice Notice HTML.
+	 * @param string $message_body original message content.
+	 * @param string $message_code message code.
+	 * @param array  $message_args associative array of message arguments.
 	 */
-	public static function notice_html( $notice ) {
+	public static function notice_html( $notice, $message_body, $message_code, $message_args ) {
 		// If the gate is not available, don't mess with the notice.
 		if ( ! self::has_gate() ) {
 			return $notice;
+		}
+		// Don't show gate unless attached to a specific post.
+		if ( empty( $message_args['post'] ) ) {
+			return '';
 		}
 		// If rendering the content in a loop, don't render the gate.
 		if ( get_queried_object_id() !== get_the_ID() ) {
@@ -310,6 +375,8 @@ class WC_Memberships {
 
 		$content = $post->post_content;
 
+		$style = \get_post_meta( $gate_post_id, 'style', true );
+
 		$use_more_tag = get_post_meta( $gate_post_id, 'use_more_tag', true );
 		// Use <!--more--> as threshold if it exists.
 		if ( $use_more_tag && strpos( $content, '<!--more-->' ) ) {
@@ -320,19 +387,58 @@ class WC_Memberships {
 			$content = explode( '</p>', $content );
 			// Extract the first $x paragraphs only.
 			$content = array_slice( $content, 0, $count ?? 2 );
+			if ( 'overlay' === $style ) {
+				// Append ellipsis to the last paragraph.
+				$content[ count( $content ) - 1 ] .= ' [&hellip;]';
+			}
 			// Rejoin the paragraphs into a single string again.
-			$content = wp_kses_post( implode( '</p>', $content ) );
+			$content = wp_kses_post( implode( '</p>', $content ) . '</p>' );
 		}
 
 		$excerpt = $content;
 
-		$style       = \get_post_meta( $gate_post_id, 'style', true );
 		$inline_fade = \get_post_meta( $gate_post_id, 'inline_fade', true );
 		if ( 'inline' === $style && $inline_fade ) {
 			$excerpt .= '<div style="pointer-events: none; height: 10em; margin-top: -10em; width: 100%; position: absolute; background: linear-gradient(180deg, rgba(255,255,255,0) 14%, rgba(255,255,255,1) 76%);"></div>';
 		}
 
 		return $excerpt;
+	}
+
+	/**
+	 * Render the overlay gate.
+	 */
+	public static function render_overlay_gate() {
+		if ( ! self::has_gate() ) {
+			return;
+		}
+		// Only render overlay gate for a restricted singular content.
+		if ( ! is_singular() || ! self::is_post_restricted() ) {
+			return;
+		}
+		$gate_post_id = self::get_gate_post_id();
+		$style        = \get_post_meta( $gate_post_id, 'style', true );
+		if ( 'overlay' !== $style ) {
+			return;
+		}
+		global $post;
+		$_post = $post;
+		$post  = \get_post( $gate_post_id ); // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
+		setup_postdata( $post );
+		$position = \get_post_meta( $gate_post_id, 'overlay_position', true );
+		$size     = \get_post_meta( $gate_post_id, 'overlay_size', true );
+		?>
+		<div class="newspack-memberships__overlay-gate" style="display:none;" data-position="<?php echo \esc_attr( $position ); ?>" data-size="<?php echo \esc_attr( $size ); ?>">
+			<div class="newspack-memberships__overlay-gate__container">
+				<div class="newspack-memberships__overlay-gate__content">
+					<?php echo \apply_filters( 'the_content', \get_the_content( null, null, $gate_post_id ) );  // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
+				</div>
+			</div>
+		</div>
+		<?php
+		self::$gate_rendered = true;
+		wp_reset_postdata();
+		$post = $_post; // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 	}
 
 	/**